refactor(svelte-app): extract profile avatar into standalone component
Moves the avatar markup, fallback handling, currentProfile subscription
and avatar styles out of PublicKeyDisplay into a dedicated ProfileAvatar
component. PublicKeyDisplay now just renders <ProfileAvatar>, keeping its
own concern limited to the public key / npub display.
